refactor: rename slide component and improve styling
