+++ src/components/ReviewSlider/ReviewCard.scss
@@ -1,0 +1,75 @@
+.review-card {
+  background-color: #ffff;
+  
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+
+  width: 100%;
+  height: 10rem;
+
+  position: relative;
+  padding: 2rem 3rem; // space for the quotes
+
+  &::before,
+  &::after {
+    content: "";
+    position: absolute;
+    width: 55px;
+    height: 40px;
+    background-size: contain;
+    background-repeat: no-repeat;
+  }
+
+  &::before {
+    top: 2rem;
+    left: 1rem;
+    background-image: url("../../images/ReviewSlider/quotes-open-icon.svg");
+  }
+
+  &::after {
+    bottom: 1.6rem;
+    right: 2rem;
+    background-image: url("../../images/ReviewSlider/quotes-close-icon.svg");
+  }
+}
+
+
+.review-card__quote {
+  margin-bottom: 0rem; // Workaround: overwrites bootstrap styling 
+  padding-left: 2.5rem;
+
+  font-size: 1.12rem;
+  font-weight: 300;
+  line-height: 24px;
+}
+
+.review-card__author {
+  padding-right: 3rem;
+
+  display: flex;
+  justify-content: end;
+  align-items: flex-end;
+}
+
+.review-card__hr {
+  display: inline-block;
+  height: 1px;
+  width: 22px;
+  background-color: #14202b;
+  border: none;
+}
+
+.review-card__name {
+  font-size: 1.12rem;
+  font-weight: 300;
+  text-transform: uppercase;
+  padding-left: 7px;
+}
+
+
+@media (min-width: 60em) {
+  .review-card {
+    width: 25rem;
+  }
+}
